refactor: created a new appendix with animations for improper integrals.
This new appendix can be swapped with the simpler appendix.tex file.
I've decided to add these animations to make the convergence and divergence of two particular improper integrals clearer. However they do not work on most pdf viewers, especially browser-based ones.
I've tested it on Okular, Evince and Firefox and it only works on Okular. Probably it will work on Adobe Acrobat as well.
Since animations are not widely supported and since their presence slows down compilation, I've decided to remove them by default, but to provide the file for those who are interested.'